// frontend/src/backend.d.ts
export interface Location {
    country: string;
    city: string;
    state: string;
}
export interface DocumentType {
    name: string;
    description: string;
}
export interface VendorRequest {
    id: string;
    status: RequestStatus;
    requester: Principal;
    contact: ContactInfo;
    name: string;
    category: string;
    products: Array<string>;
    location: Location;
}
// ...
export interface ContactInfo {
    email: string;
    website: string;
    address: string;
    phone: string;
}
// ...
export enum RequestStatus {
    pending = "pending",
    approved = "approved",
    rejected = "rejected"
}